Parallelize span ingestion
This adds a num_span_workers config value and makes an array of span
workers which all process spans in parallel on the defined span sinks.

Much like ingestion for metrics, the span worker that gets the span is
decided by a pseudorandom number (the span ID, or if none is set, a
randomly generated value).